add entrypoint to frontend & move vue to frontend
