+++ app/javascript/packs/actnesses/index.js
@@ -1,6 +1,6 @@
 
 import Vue from 'vue'
-import ActnessIndex from '../../organisms/actnesses/index.vue'
+import ActnessIndex from '../../../../frontend/javascript/components/organisms/actnesses/index.vue'
 
 document.addEventListener('DOMContentLoaded', () => {
     const el = document.body.appendChild(document.createElement('hello'))
