+++ app/javascript/packs/home/index.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import HomeIndex from '../../organisms/home/index.vue'
+import HomeIndex from '../../../../frontend/javascript/components/organisms/home/index.vue'
 
 document.addEventListener('DOMContentLoaded', () => {
     const el = document.body.appendChild(document.createElement('hello'))
