+++ app/javascript/packs/runnings/index.js
@@ -1,6 +1,6 @@
 
 import Vue from 'vue'
-import RunningIndex from '../../organisms/runnings/index.vue'
+import RunningIndex from '../../../../frontend/javascript/components/organisms/runnings/index.vue'
 
 document.addEventListener('DOMContentLoaded', () => {
     const el = document.body.appendChild(document.createElement('hello'))
