+++ app/views/actnesses/index.html.erb
@@ -39,4 +39,5 @@
 <%= link_to 'Back_to_Situation', situations_path %>|
 <%= link_to 'Go_to_Running', runnings_path %>
 
-<%= javascript_pack_tag 'actnesses/index' %>
+<%= javascript_pack_tag 'actnesses' %>
+<div data-vue="ActnessesIndex"></div>
