+++ app/views/games/index.html.erb
@@ -31,4 +31,5 @@
 <%= link_to 'Home', home_index_path %>|
 <%= link_to 'Go_to_Situation', situations_path %>
 
-<%= javascript_pack_tag 'games/index' %>
+<%= javascript_pack_tag 'games' %>
+<div data-vue="GamesIndex"></div>
