+++ app/views/runnings/index.html.erb
@@ -35,4 +35,5 @@
 <%= link_to 'Back_to_Batting', actnesses_path %>|
 <%= link_to 'Home', home_index_path %>
 
-<%= javascript_pack_tag 'runnings/index' %>
+<%= javascript_pack_tag 'runnings' %>
+<div data-vue="RunningsIndex"></div>
