+++ app/views/situations/index.html.erb
@@ -45,4 +45,5 @@
 <%= link_to 'Back_to_Game', games_path %>|
 <%= link_to 'Go_to_actness', actnesses_path %>
 
-<%= javascript_pack_tag 'situations/index' %>
+<%= javascript_pack_tag 'situations' %>
+<div data-vue="SituationsIndex"></div>
