+++ frontend/javascript/components/organisms/games/index.vue
