+++ frontend/javascript/components/organisms/situations/index.vue
