+++ frontend/javascript/entry/actnesses.js
@@ -1,8 +1,8 @@
-import App from "../app.vue";
+import ActnessesIndex from '../components/organisms/actnesses/index.vue'
 import * as VueMounter from "../application.js"
 
 export const components = {
-    App
+    ActnessesIndex
 };
 
 VueMounter.mount(components);
